+++ demisto_sdk/__main__.py
@@ -850,7 +850,7 @@ def update_pack_releasenotes(**kwargs):
             print_warning('No changes were detected. If changes were made, please commit the changes '
                           'and rerun the command')
         else:
-            update_pack_rn = UpdateRN(pack_path=_pack, update_type=update_type, pack_files=set(),
+            update_pack_rn = UpdateRN(pack_path=_pack, update_type=update_type, modified_files_in_pack=set(),
                                       pre_release=pre_release, added_files=set(),
                                       specific_version=specific_version, pack_metadata_only=True)
             update_pack_rn.execute_update()
@@ -861,9 +861,9 @@ def update_pack_releasenotes(**kwargs):
         packs_list = ''.join(f"{p}, " for p in packs)
         print_warning(f"Adding release notes to the following packs: {packs_list.rstrip(', ')}")
         for pack in packs:
-            update_pack_rn = UpdateRN(pack_path=pack, update_type=update_type, pack_files=modified.union(old),
-                                      pre_release=pre_release, added_files=added,
-                                      specific_version=specific_version)
+            update_pack_rn = UpdateRN(pack_path=pack, update_type=update_type,
+                                      modified_files_in_pack=modified.union(old), pre_release=pre_release,
+                                      added_files=added, specific_version=specific_version)
             update_pack_rn.execute_update()
     elif is_all and _pack:
         print_error("Please remove the --all flag when specifying only one pack.")
@@ -880,9 +880,9 @@ def update_pack_releasenotes(**kwargs):
                             f"Please update manually or run `demisto-sdk update-release-notes "
                             f"-p {_pack}` without specifying the update_type.")
             else:
-                update_pack_rn = UpdateRN(pack_path=_pack, update_type=update_type, pack_files=modified.union(old),
-                                          pre_release=pre_release, added_files=added,
-                                          specific_version=specific_version)
+                update_pack_rn = UpdateRN(pack_path=_pack, update_type=update_type,
+                                          modified_files_in_pack=modified.union(old), pre_release=pre_release,
+                                          added_files=added, specific_version=specific_version)
                 update_pack_rn.execute_update()