+++ demisto_sdk/commands/common/hook_validations/layout.py
@@ -1,3 +1,4 @@
+import os
 from abc import ABC, abstractmethod
 from distutils.version import LooseVersion
 
@@ -20,11 +21,13 @@ class LayoutBaseValidator(ContentEntityValidator, ABC):
         Returns:
             bool. Whether the layout is valid or not
         """
-        return (super().is_valid_file(validate_rn) and
-                self.is_valid_version() and
-                self.is_valid_from_version() and
-                self.is_valid_to_version() and
-                self.is_to_version_higher_than_from_version())
+        return all([super().is_valid_file(validate_rn),
+                    self.is_valid_version(),
+                    self.is_valid_from_version(),
+                    self.is_valid_to_version(),
+                    self.is_to_version_higher_than_from_version(),
+                    self.is_valid_file_path()
+                    ])
 
     def is_valid_version(self) -> bool:
         """Checks if version field is valid. uses default method.
@@ -55,6 +58,10 @@ class LayoutBaseValidator(ContentEntityValidator, ABC):
     def is_valid_to_version(self) -> bool:
         pass
 
+    @abstractmethod
+    def is_valid_file_path(self) -> bool:
+        pass
+
 
 class LayoutsContainerValidator(LayoutBaseValidator):
     def is_valid_from_version(self) -> bool:
@@ -77,6 +84,14 @@ class LayoutsContainerValidator(LayoutBaseValidator):
         """
         if self.to_version and LooseVersion(self.to_version) < LooseVersion(FROM_VERSION_LAYOUTS_CONTAINER):
             error_message, error_code = Errors.invalid_version_in_layoutscontainer('toVersion')
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                return False
+        return True
+
+    def is_valid_file_path(self) -> bool:
+        output_basename = os.path.basename(self.file_path)
+        if not output_basename.startswith('layoutscontainer-'):
+            error_message, error_code = Errors.invalid_file_path_layoutscontainer(output_basename)
             if self.handle_error(error_message, error_code, file_path=self.file_path):
                 return False
         return True
@@ -105,6 +120,14 @@ class LayoutValidator(LayoutBaseValidator):
         """
         if not self.to_version or LooseVersion(self.to_version) >= LooseVersion(FROM_VERSION_LAYOUTS_CONTAINER):
             error_message, error_code = Errors.invalid_version_in_layout('toVersion')
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                return False
+        return True
+
+    def is_valid_file_path(self) -> bool:
+        output_basename = os.path.basename(self.file_path)
+        if not output_basename.startswith('layout-'):
+            error_message, error_code = Errors.invalid_file_path_layout(output_basename)
             if self.handle_error(error_message, error_code, file_path=self.file_path):
                 return False
         return True