+++ demisto_sdk/commands/update_release_notes/tests/update_rn_test.py
@@ -34,9 +34,8 @@ class TestRNUpdate(unittest.TestCase):
             "\n#### Scripts\n##### Hello World Script\n- %%UPDATE_RN%%\n" \
             "\n#### Widgets\n##### Hello World Widget\n- %%UPDATE_RN%%\n"
 
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         changed_items = {
             "Hello World Integration": {"type": "Integration", "description": "", "is_new_file": False},
             "Hello World Playbook": {"type": "Playbook", "description": "", "is_new_file": False},
@@ -69,9 +68,8 @@ class TestRNUpdate(unittest.TestCase):
         expected_result = "\n#### Playbooks\n##### New: Hello World Playbook\n- Hello World Playbook description\n" \
 
 
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         changed_items = {
             "Hello World Playbook": {"type": "Playbook",
                                      "description": "Hello World Playbook description", "is_new_file": True},
@@ -91,9 +89,8 @@ class TestRNUpdate(unittest.TestCase):
         expected_result = "\n#### Playbooks\n##### Hello World Playbook\n- %%UPDATE_RN%%\n" \
 
 
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         changed_items = {
             "Hello World Playbook": {"type": "Playbook",
                                      "description": "Hello World Playbook description", "is_new_file": False},
@@ -114,9 +111,8 @@ class TestRNUpdate(unittest.TestCase):
         expected_result = "\n#### Incident Fields\n- **Hello World IncidentField**\n" \
 
 
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         changed_items = {
             "Hello World IncidentField": {"type": "Incident Fields", "description": "", "is_new_file": False},
         }
@@ -137,7 +133,7 @@ class TestRNUpdate(unittest.TestCase):
                           "- Maintenance and stability enhancements.\n"\
 
         from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='maintenance', pack_files={'HelloWorld'},
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='maintenance', modified_files_in_pack={'HelloWorld'},
                              added_files=set())
         changed_items = {
             "Hello World Integration": {"type": "Integration", "description": "", "is_new_file": False},
@@ -159,7 +155,7 @@ class TestRNUpdate(unittest.TestCase):
                           "- Documentation and metadata improvements.\n"\
 
         from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='documentation', pack_files={'HelloWorld'},
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='documentation', modified_files_in_pack={'HelloWorld'},
                              added_files=set())
         changed_items = {
             "Hello World Integration": {"type": "Integration", "description": "", "is_new_file": False},
@@ -179,9 +175,8 @@ class TestRNUpdate(unittest.TestCase):
         expected_result = "\n#### Integrations\n##### HelloWorld\n- Documentation and metadata improvements.\n"\
 
 
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files=set(), added_files=set(),
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack=set(), added_files=set(),
                              pack_metadata_only=True)
         changed_items = {}
         release_notes = update_rn.build_rn_template(changed_items)
@@ -202,21 +197,21 @@ class TestRNUpdate(unittest.TestCase):
             UpdateRN
 
         # case 1:
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld/README.md'},
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld/README.md'},
                              added_files=set())
         assert update_rn.only_readme_changed()
 
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files=set(),
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack=set(),
                              added_files={'HelloWorld/README.md'})
         assert update_rn.only_readme_changed()
 
         # case 2:
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld/README.md'},
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld/README.md'},
                              added_files={'HelloWorld/HelloWorld.py'})
         assert not update_rn.only_readme_changed()
 
         update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor',
-                             pack_files={'HelloWorld/HelloWorld.yml', 'HelloWorld/README.md'}, added_files=set())
+                             modified_files_in_pack={'HelloWorld/HelloWorld.yml', 'HelloWorld/README.md'}, added_files=set())
         assert not update_rn.only_readme_changed()
 
     def test_find_corresponding_yml(self):
@@ -229,9 +224,8 @@ class TestRNUpdate(unittest.TestCase):
                 - return only the yml of the changed file
         """
         expected_result = "Integration/HelloWorld.yml"
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         filepath = 'Integration/HelloWorld.py'
         filename = update_rn.find_corresponding_yml(filepath)
         assert expected_result == filename
@@ -246,9 +240,8 @@ class TestRNUpdate(unittest.TestCase):
                 - the filepath of the correct release notes.
         """
         expected_result = 'Packs/HelloWorld/ReleaseNotes/1_1_1.md'
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         input_version = '1.1.1'
         result = update_rn.return_release_notes_path(input_version)
         assert expected_result == result
@@ -265,9 +258,8 @@ class TestRNUpdate(unittest.TestCase):
         shutil.copy(src=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack/pack_metadata.json'),
                     dst=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack/_pack_metadata.json'))
         expected_version = '1.1.0'
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.metadata_path = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack/pack_metadata.json')
         version_number, _ = update_rn.bump_version_number(pre_release=False, specific_version=None)
         assert version_number == expected_version
@@ -287,9 +279,8 @@ class TestRNUpdate(unittest.TestCase):
         shutil.copy(src=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack/pack_metadata.json'),
                     dst=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack/_pack_metadata.json'))
         expected_version = '2.0.0'
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='major', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='major', modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.metadata_path = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack/pack_metadata.json')
         version_number, _ = update_rn.bump_version_number(pre_release=False, specific_version=None)
         assert version_number == expected_version
@@ -309,9 +300,8 @@ class TestRNUpdate(unittest.TestCase):
         shutil.copy(src=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack/pack_metadata.json'),
                     dst=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack/_pack_metadata.json'))
         expected_version = '1.0.1'
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='revision', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='revision', modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.metadata_path = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack/pack_metadata.json')
         version_number, _ = update_rn.bump_version_number(pre_release=False, specific_version=None)
         assert version_number == expected_version
@@ -334,7 +324,7 @@ class TestRNUpdate(unittest.TestCase):
         from demisto_sdk.commands.update_release_notes.update_rn import \
             UpdateRN
         update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type=None, specific_version='2.0.0',
-                             pack_files={'HelloWorld'}, added_files=set())
+                             modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.metadata_path = os.path.join(TestRNUpdate.FILES_PATH,
                                                'fake_pack/pack_metadata.json')
         version_number, _ = update_rn.bump_version_number(pre_release=False,
@@ -355,9 +345,8 @@ class TestRNUpdate(unittest.TestCase):
         """
         shutil.copy(src=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json'),
                     dst=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/_pack_metadata.json'))
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='revision', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='revision', modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.metadata_path = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json')
         self.assertRaises(ValueError, update_rn.bump_version_number)
         os.remove(os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json'))
@@ -375,9 +364,8 @@ class TestRNUpdate(unittest.TestCase):
         """
         shutil.copy(src=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json'),
                     dst=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/_pack_metadata.json'))
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.metadata_path = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json')
         self.assertRaises(ValueError, update_rn.bump_version_number)
         os.remove(os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json'))
@@ -395,9 +383,8 @@ class TestRNUpdate(unittest.TestCase):
         """
         shutil.copy(src=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json'),
                     dst=os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/_pack_metadata.json'))
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='major', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='major', modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.metadata_path = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json')
         self.assertRaises(ValueError, update_rn.bump_version_number)
         os.remove(os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json'))
@@ -413,9 +400,8 @@ class TestRNUpdate(unittest.TestCase):
             Then:
                 - return ValueError
         """
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='major', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='major', modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.metadata_path = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata_.json')
         self.assertRaises(SystemExit, update_rn.bump_version_number)
 
@@ -428,9 +414,8 @@ class TestRNUpdate(unittest.TestCase):
             Then:
                 - return ValueError
         """
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type=None, pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type=None, modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.metadata_path = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json')
         self.assertRaises(ValueError, update_rn.bump_version_number)
 
@@ -499,9 +484,8 @@ class TestRNUpdateUnit:
             Then:
                 - return tuple where first value is the pack name, and second is the item type
         """
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path=pack_name, update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path=pack_name, update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         filepath = os.path.join(TestRNUpdate.FILES_PATH, path)
         mocker.patch.object(UpdateRN, 'find_corresponding_yml', return_value='Integrations/VulnDB/VulnDB.yml')
         mocker.patch.object(UpdateRN, 'get_display_name', return_value='VulnDB')
@@ -520,7 +504,7 @@ class TestRNUpdateUnit:
         from demisto_sdk.commands.update_release_notes.update_rn import \
             UpdateRN
         filepath = os.path.join(TestRNUpdate.FILES_PATH, 'ReleaseNotes')
-        update_rn = UpdateRN(pack_path="Packs/VulnDB", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        update_rn = UpdateRN(pack_path="Packs/VulnDB", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         update_rn.check_rn_dir(filepath)
 
     def test_create_markdown(self):
@@ -532,9 +516,8 @@ class TestRNUpdateUnit:
             Then:
                 - create the file or skip if it exists.
         """
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/VulnDB", update_type='minor', pack_files={'HelloWorld'}, added_files=set())
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/VulnDB", update_type='minor', modified_files_in_pack={'HelloWorld'}, added_files=set())
         filepath = os.path.join(TestRNUpdate.FILES_PATH, 'ReleaseNotes/1_1_1.md')
         md_string = '### Test'
         update_rn.create_markdown(release_notes_path=filepath, rn_string=md_string, changed_files={})
@@ -548,9 +531,8 @@ class TestRNUpdateUnit:
             Then:
                 - return updated release notes while preserving the integrity of the existing notes.
         """
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'},
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'},
                              added_files=set())
         new_rn = update_rn.update_existing_rn(self.CURRENT_RN, self.CHANGED_FILES)
         assert self.EXPECTED_RN_RES == new_rn
@@ -566,9 +548,8 @@ class TestRNUpdateUnit:
         """
         ORIGINAL = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/pack_metadata.json')
         TEMP_FILE = os.path.join(TestRNUpdate.FILES_PATH, 'fake_pack_invalid/_pack_metadata.json')
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files={'HelloWorld'},
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack={'HelloWorld'},
                              added_files=set())
         shutil.copy(src=ORIGINAL, dst=TEMP_FILE)
         data_dict = get_json(TEMP_FILE)
@@ -589,10 +570,10 @@ class TestRNUpdateUnit:
         from demisto_sdk.commands.update_release_notes.update_rn import \
             UpdateRN
         added_files = {'HelloWorld/something_new.md', 'HelloWorld/test_data/nothing.md'}
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files=set(),
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack=set(),
                              added_files=added_files)
         update_rn.find_added_pack_files()
-        assert update_rn.pack_files == {'HelloWorld/something_new.md'}
+        assert update_rn.modified_files_in_pack == {'HelloWorld/something_new.md'}
 
     def test_does_pack_metadata_exist_no(self):
         """
@@ -603,9 +584,8 @@ class TestRNUpdateUnit:
             Then:
                 - return False to indicate it does not exist.
         """
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', pack_files=set(),
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/HelloWorld", update_type='minor', modified_files_in_pack=set(),
                              added_files=set())
         update_rn.metadata_path = 'This/Doesnt/Exist'
         result = update_rn._does_pack_metadata_exist()
@@ -620,9 +600,8 @@ class TestRNUpdateUnit:
             Then:
                 - return an error message and exit.
         """
-        from demisto_sdk.commands.update_release_notes.update_rn import \
-            UpdateRN
-        update_rn = UpdateRN(pack_path="Packs/Legacy", update_type='minor', pack_files=set(),
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        update_rn = UpdateRN(pack_path="Packs/Legacy", update_type='minor', modified_files_in_pack=set(),
                              added_files=set())
         update_rn.execute_update()
 
@@ -658,7 +637,7 @@ class TestRNUpdateUnit:
         from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
         from subprocess import Popen
         import json
-        update_rn = UpdateRN(pack_path="Packs/Base", update_type='minor', pack_files=set(),
+        update_rn = UpdateRN(pack_path="Packs/Base", update_type='minor', modified_files_in_pack=set(),
                              added_files=set())
         mocker.patch.object(UpdateRN, 'get_pack_metadata',
                             return_value={"currentVersion": pack_current_version})
@@ -683,10 +662,31 @@ class TestRNUpdateUnit:
             ('file2', 'file2_new'),
             'file3'
         }
-        update_rn = UpdateRN(pack_path="Packs/Base", update_type='minor', pack_files=modified_files,
+        update_rn = UpdateRN(pack_path="Packs/Base", update_type='minor', modified_files_in_pack=modified_files,
                              added_files=set())
 
-        assert 'file1' in update_rn.pack_files
-        assert 'file2_new' in update_rn.pack_files
-        assert ('file2', 'file2_new') not in update_rn.pack_files
-        assert 'file3' in update_rn.pack_files
+        assert 'file1' in update_rn.modified_files_in_pack
+        assert 'file2_new' in update_rn.modified_files_in_pack
+        assert ('file2', 'file2_new') not in update_rn.modified_files_in_pack
+        assert 'file3' in update_rn.modified_files_in_pack
+
+    def test_check_for_release_notes_valid_file_path(self):
+        """
+        Given:
+            case 1: a description file
+            case 2: an image file
+            case 3: a non-image or description file
+        When:
+            running check_for_release_notes_valid_file_path method
+        Then:
+            case 1 & 2: change the file path to the corresponding yml file.
+            case 3: file path remains unchnaged
+        """
+        from demisto_sdk.commands.update_release_notes.update_rn import UpdateRN
+        image_file_path = "Packs/DNSDB/Integrations/DNSDB_v2/DNSDB_v2_image.png"
+        description_file_path = "Packs/DNSDB/Integrations/DNSDB_v2/DNSDB_v2_description.md"
+        yml_file_path = "Packs/DNSDB/Integrations/DNSDB_v2/DNSDB_v2.yml"
+
+        assert yml_file_path == UpdateRN.check_for_release_notes_valid_file_path(image_file_path)
+        assert yml_file_path == UpdateRN.check_for_release_notes_valid_file_path(description_file_path)
+        assert yml_file_path == UpdateRN.check_for_release_notes_valid_file_path(yml_file_path)