+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 * Fixed an issue where **validate** validated all images instead of packs only.
 * Fixed an issue where a warning was not printed in the **format** in case a non-supported file type is inputted.
 * Fixed an issue where **validate** did not fail if no release notes were added when adding files to existing packs.
+* Added handling of incorrect layout paths via the **format** command.
 * Refactor **create-content-artifacts** command - Efficient artifacts creation and better logging.
+* Fixed an issue where image and description files were not handled correctly by **validate** and **update-release-notes** commands.
 
 # 1.1.11
 * Added line number to secrets' path in **secrets** command report.